# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Akmatori API
  description: |
    AI-powered AIOps platform API. Manages skills, tools, incidents, alert sources,
    and system settings for automated incident investigation and remediation.
  version: 1.0.0
  contact:
    name: Akmatori
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /auth/login:
    post:
      summary: Login
      operationId: login
      security: []
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  username:
                    type: string
                  expires_in:
                    type: integer
                    description: Token expiry in seconds
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/verify:
    get:
      summary: Verify token
      operationId: verifyToken
      tags: [Authentication]
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  username:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===== Skills =====
...